OTC-1080: added check for pssql to delete dependent table before deleting core_user

+++ core/test_services.py
@@ -1,5 +1,7 @@
 import importlib
 import logging
+
+from django.db import connection
 from django.test.client import RequestFactory
 from django.apps import apps
 
@@ -470,7 +472,8 @@ class UserServicesTest(TestCase):
         self.assertTrue(mail.outbox[0].subject == "[OpenIMIS] Reset Password")
 
         UserRole.objects.filter(user_id=i_user.id).delete()
-        UserRole.objects.filter(user_id=core_user.id).delete()
+        if connection.vendor == "postgresql":
+            UserRole.objects.filter(user_id=core_user.id).delete()
         core_user.delete()
         i_user.delete()
 
@@ -505,6 +508,7 @@ class UserServicesTest(TestCase):
             set_user_password(request, username, "TOKEN", "new_password")
 
         UserRole.objects.filter(user_id=i_user.id).delete()
-        UserRole.objects.filter(user_id=core_user.id).delete()
+        if connection.vendor == "postgresql":
+            UserRole.objects.filter(user_id=core_user.id).delete()
         core_user.delete()
         i_user.delete()
